Add support for pylint 1.5.2
Apart of incrementing the pylint version, the changeset involved
modifying the way in which the messages were parsed from pylint.
Previously they were formatted with a hand crafted message template,
which resulted in potentially invalid JSON, especially for messages
which put additional quotes into the message (for instance
expression-not-assigned). This was dropped in favor of building
manually the issue class from the parsed message lines.
The test files were also updated so that they could work on
the new release. Other things which were updated were removed
message ids or message ids which now have a different value,
as it is the case for E0235, which is now E0302 (unexpected-special-method-signature).

(FT-1054)