Support analyzing files with Python 3 specific syntax
Python 3 is not always backward compatible with Python 2,
due to specific syntax that's found only in the former.
Pylint, unfortunately, is tied up to the current interpreter's
AST library, which is not backward compatible. This means
that files with Python 3 or Python 2 specific syntax needs
to be analyzed with a specific installation of Pylint.
In order to support both versions, we're partitioning
the list of files by their required Python version.
The classification is done by pre-parsing each file with
Python's ast module and figure out which interpreter it might
need (if it raises a SyntaxError, then it's a high possibility
of needing the other version of Python). The results are
then propagating together as if only one program executed,
not two.

(FT-970)